+++ test/terra/noise/passes/test_local_noise_pass.py
@@ -10,19 +10,16 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """
-Noise pass classes tests
+LocalNoisePass class tests
 """
 
-from qiskit.providers.aer.noise.passes import LocalNoisePass
-from qiskit.providers.aer.noise.passes import RelaxationNoisePass
 from qiskit.providers.aer.noise.errors import ReadoutError
-from test.terra.common import QiskitAerTestCase
+from qiskit.providers.aer.noise.passes import LocalNoisePass
 
-from qiskit.circuit import QuantumCircuit, Delay
+from qiskit.circuit import QuantumCircuit
 from qiskit.circuit.library.standard_gates import SXGate, HGate
-from qiskit.compiler import transpile
-from qiskit.test.mock import FakeLagos
-from qiskit.transpiler import InstructionDurations, TranspilerError
+from qiskit.transpiler import TranspilerError
+from test.terra.common import QiskitAerTestCase
 
 
 class TestLocalNoisePass(QiskitAerTestCase):
@@ -98,55 +95,3 @@ class TestLocalNoisePass(QiskitAerTestCase):
         noise_pass = LocalNoisePass(func=out_readout_error, op_types=HGate)
         with self.assertRaises(TranspilerError):
             noise_pass(qc)
-
-
-class TestRelaxationNoisePass(QiskitAerTestCase):
-    """Testing RelaxationNoisePass class"""
-
-    def test_default_with_scheduled_circuit(self):
-        """Test adding noises to all ops in a scheduled circuit."""
-        qc = QuantumCircuit(2, 2)
-        qc.h(0)
-        qc.cx(0, 1)
-        qc.measure([0, 1], [0, 1])
-
-        backend = FakeLagos()
-
-        sched_circ = transpile(qc, backend, scheduling_method='alap')
-
-        noise_pass = RelaxationNoisePass(
-            t1s=[backend.properties().t1(q) for q in range(backend.configuration().num_qubits)],
-            t2s=[backend.properties().t2(q) for q in range(backend.configuration().num_qubits)],
-            instruction_durations=InstructionDurations.from_backend(backend)
-        )
-        noisy_circ = noise_pass(sched_circ)
-        self.assertEqual(9, noisy_circ.count_ops()["quantum_channel"])
-
-    def test_raise_if_supplied_invalid_ops(self):
-        with self.assertRaises(TranspilerError):
-            RelaxationNoisePass(
-                t1s=[1],
-                t2s=[1],
-                instruction_durations=InstructionDurations(),
-                op_types="delay",  # str is invalid
-            )
-
-    def test_ops_option_with_scheduled_circuit(self):
-        """Test adding noises only to delays in a scheduled circuit."""
-        qc = QuantumCircuit(2, 2)
-        qc.h(0)
-        qc.cx(0, 1)
-        qc.measure([0, 1], [0, 1])
-
-        backend = FakeLagos()
-
-        sched_circ = transpile(qc, backend, scheduling_method='alap')
-
-        delay_pass = RelaxationNoisePass(
-            t1s=[backend.properties().t1(q) for q in range(backend.configuration().num_qubits)],
-            t2s=[backend.properties().t2(q) for q in range(backend.configuration().num_qubits)],
-            instruction_durations=InstructionDurations.from_backend(backend),
-            op_types=Delay,
-        )
-        noisy_circ = delay_pass(sched_circ)
-        self.assertEqual(6, noisy_circ.count_ops()["quantum_channel"])
